+++ module_usb_audio/audio_io/userbuffermanagement.c
@@ -1,14 +1,17 @@
-
+#include "userbuffermanagement.h"
 #include "xccompat.h"
 #include "devicedefines.h"
 
-/* Default implentation for UserBufferManagement() */
-void UserBufferManagement(unsigned sampsFromUsbToAudio[], unsigned sampsFromAudioToUsb[]
-        , unsigned i_audMan
-        ) __attribute__ ((weak));
-void UserBufferManagement(unsigned sampsFromUsbToAudio[], unsigned sampsFromAudioToUsb[]
-        , unsigned i_audMan
-)
+/* Default implementation for UserBufferManagementInit() */
+void __attribute__ ((weak)) UserBufferManagementInit(CLIENT_INTERFACE(audManage_if, i_audMan))
+{
+    /* Do nothing */
+}
+
+/* Default implementation for UserBufferManagement() */
+void __attribute__ ((weak)) UserBufferManagement(unsigned sampsFromUsbToAudio[],
+                                                 unsigned sampsFromAudioToUsb[],
+                                                 CLIENT_INTERFACE(audManage_if, i_audMan))
 {
     /* Do nothing */
 }